requirements.txt
numpy
pandas
scikit-learn
python-dotenv
pyyaml
pymongo
certifi
pymongo[srv]==3.11

# -e .